backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo==4.5.0
python-dotenv==1.0.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bson==0.5.10 